moved scc constants in .constants module
